port(filament): move jsMain actuals to webMain; js target green
- web: adopt kotlinx-browser (0.5.0) and remap the typed-array/canvas externals
  back to org.khronos.webgl / org.w3c.dom so the DOM + typed-array actuals compile
  unchanged on both js and wasmJs (avoids rewriting them to kotlin-wrappers APIs).
- filament: jsMain -> webMain; add .toDouble() at external call sites (dropped the
  Double->Number remap); add else branches to the enum-mapping `when`s (external
  enums are no longer sealed after the wasmJs lowering).

:kotlin:filament compiles for js. wasmJs still needs the JS-array/matrix marshaling
layer (Array<Number> -> JsArray, top-level js(), unsafeCast rework) — follow-up.

+++ kotlin/filament/src/webMain/kotlin/io/github/erkko68/filament/BufferObject.js.kt
@@ -24,7 +24,7 @@ actual class BufferObject(internal val jsBufferObject: JSBufferObject) {
         count: Int
     ) {
         val clippedData = if (count < data.size) data.sliceArray(0 until count) else data
-        jsBufferObject.setBuffer(engine.jsEngine, clippedData.toUint8Array(), destOffsetInBytes)
+        jsBufferObject.setBuffer(engine.jsEngine, clippedData.toUint8Array(), destOffsetInBytes.toDouble())
     }
 
     actual fun setBuffer(
@@ -43,7 +43,7 @@ actual class BufferObject(internal val jsBufferObject: JSBufferObject) {
         private val jsBuilder = JSBufferObject.Builder()
 
         actual fun size(byteCount: Int): Builder {
-            jsBuilder.size(byteCount)
+            jsBuilder.size(byteCount.toDouble())
             return this
         }
