+++ kotlin/filament/src/webMain/kotlin/io/github/erkko68/filament/TextureSampler.js.kt
@@ -62,7 +62,7 @@ actual class TextureSampler {
         val magJs = mapMagFilter(magFilter)
         val wrapJs = mapWrapMode(wrapModeS)
         val ts = JSTextureSampler(minJs.unsafeCast<JSMinFilter>(), magJs, wrapJs)
-        ts.setAnisotropy(anisotropy)
+        ts.setAnisotropy(anisotropy.toDouble())
         if (compareMode != CompareMode.NONE) {
             ts.setCompareMode(mapCompareMode(compareMode), mapCompareFunc(compareFunction))
         }
